Insert abstract class for SoftwareModule and DistributinSet TableHeader
