+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/distributions/smtable/SwModuleTableHeader.java
@@ -12,36 +12,27 @@ import org.eclipse.hawkbit.ui.SpPermissionChecker;
 import org.eclipse.hawkbit.ui.artifacts.event.SMFilterEvent;
 import org.eclipse.hawkbit.ui.artifacts.event.SoftwareModuleEvent;
 import org.eclipse.hawkbit.ui.artifacts.smtable.SoftwareModuleAddUpdateWindow;
-import org.eclipse.hawkbit.ui.common.table.AbstractTableHeader;
+import org.eclipse.hawkbit.ui.common.table.AbstractSoftwareModuleTableHeader;
 import org.eclipse.hawkbit.ui.common.table.BaseEntityEventType;
 import org.eclipse.hawkbit.ui.distributions.event.DistributionsUIEvent;
 import org.eclipse.hawkbit.ui.distributions.state.ManageDistUIState;
 import org.eclipse.hawkbit.ui.utils.I18N;
-import org.eclipse.hawkbit.ui.utils.UIComponentIdProvider;
 import org.vaadin.spring.events.EventBus.UIEventBus;
 import org.vaadin.spring.events.EventScope;
 import org.vaadin.spring.events.annotation.EventBusListenerMethod;
-
-import com.vaadin.event.dd.DropHandler;
-import com.vaadin.ui.Button.ClickEvent;
-import com.vaadin.ui.UI;
-import com.vaadin.ui.Window;
 
 /**
  * Implementation of software module Header block using generic abstract details
  * style .
  */
-public class SwModuleTableHeader extends AbstractTableHeader {
+public class SwModuleTableHeader extends AbstractSoftwareModuleTableHeader {
 
-    private static final long serialVersionUID = 242961845006626297L;
-
-    private final SoftwareModuleAddUpdateWindow softwareModuleAddUpdateWindow;
+    private static final long serialVersionUID = 1L;
 
     SwModuleTableHeader(final I18N i18n, final SpPermissionChecker permChecker, final UIEventBus eventbus,
             final ManageDistUIState manageDistUIstate,
             final SoftwareModuleAddUpdateWindow softwareModuleAddUpdateWindow) {
-        super(i18n, permChecker, eventbus, null, manageDistUIstate, null);
-        this.softwareModuleAddUpdateWindow = softwareModuleAddUpdateWindow;
+        super(i18n, permChecker, eventbus, null, manageDistUIstate, null, softwareModuleAddUpdateWindow);
     }
 
     @EventBusListenerMethod(scope = EventScope.UI)
@@ -52,50 +43,8 @@ public class SwModuleTableHeader extends AbstractTableHeader {
     }
 
     @Override
-    protected String getHeaderCaption() {
-        return i18n.get("upload.swModuleTable.header");
-    }
-
-    @Override
-    protected String getSearchBoxId() {
-        return UIComponentIdProvider.SW_MODULE_SEARCH_TEXT_FIELD;
-    }
-
-    @Override
-    protected String getSearchRestIconId() {
-        return UIComponentIdProvider.SW_MODULE_SEARCH_RESET_ICON;
-    }
-
-    @Override
-    protected String getAddIconId() {
-        return UIComponentIdProvider.SW_MODULE_ADD_BUTTON;
-    }
-
-    @Override
     protected String onLoadSearchBoxValue() {
         return manageDistUIstate.getSoftwareModuleFilters().getSearchText().orElse(null);
-    }
-
-    @Override
-    protected String getDropFilterId() {
-        /* not required */
-        return null;
-    }
-
-    @Override
-    protected boolean isDropHintRequired() {
-        return false;
-    }
-
-    @Override
-    protected boolean isDropFilterRequired() {
-        /* Not Yet Implemented */
-        return false;
-    }
-
-    @Override
-    protected String getShowFilterButtonLayoutId() {
-        return "show.type.icon";
     }
 
     @Override
@@ -110,11 +59,6 @@ public class SwModuleTableHeader extends AbstractTableHeader {
             manageDistUIstate.getSoftwareModuleFilters().setSearchText(null);
             eventbus.publish(this, SMFilterEvent.REMOVER_FILTER_BY_TEXT);
         }
-    }
-
-    @Override
-    protected String getMaxMinIconId() {
-        return UIComponentIdProvider.SW_MAX_MIN_TABLE_ICON;
     }
 
     @Override
@@ -146,56 +90,4 @@ public class SwModuleTableHeader extends AbstractTableHeader {
         eventbus.publish(this, SMFilterEvent.FILTER_BY_TEXT);
     }
 
-    @Override
-    protected void addNewItem(final ClickEvent event) {
-        final Window addSoftwareModule = softwareModuleAddUpdateWindow.createAddSoftwareModuleWindow();
-        addSoftwareModule.setCaption(i18n.get("upload.caption.add.new.swmodule"));
-        UI.getCurrent().addWindow(addSoftwareModule);
-        addSoftwareModule.setVisible(Boolean.TRUE);
-    }
-
-    @Override
-    protected boolean hasCreatePermission() {
-        return permChecker.hasCreateDistributionPermission();
-    }
-
-    @Override
-    protected Boolean isAddNewItemAllowed() {
-        return true;
-    }
-
-    @Override
-    protected String getFilterIconStyle() {
-        return null;
-    }
-
-    @Override
-    protected String getDropFilterWrapperId() {
-        return null;
-    }
-
-    @Override
-    protected DropHandler getDropFilterHandler() {
-        return null;
-    }
-
-    @Override
-    protected String getBulkUploadIconId() {
-        return null;
-    }
-
-    @Override
-    protected void bulkUpload(final ClickEvent event) {
-        // No implementation as no bulk upload is supported.
-    }
-
-    @Override
-    protected Boolean isBulkUploadAllowed() {
-        return Boolean.FALSE;
-    }
-
-    @Override
-    protected boolean isBulkUploadInProgress() {
-        return false;
-    }
 }
